backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django==4.2.10",
    "django-cors-headers==4.3.1",
    "djangorestframework==3.14.0",
    "pillow==10.1.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv==1.0.0",
    "qrcode==7.4.2",
    "requests>=2.32.3",
    "squareup>=42.1.0.20250416",
    "stripe>=12.1.0",
]
